refactor[#13] : 파일 구조 변경 및 테스트 완료
