+++ fastAPI/NICKNAME-GENERATION-API/.gitignore
@@ -1,0 +1,4 @@
+.env
+.venv
+__pycache__
+logs
