+++ fastAPI/NICKNAME-GENERATION-API/dockerfile
@@ -1,0 +1,9 @@
+FROM public.ecr.aws/lambda/python:3.9
+
+COPY requirements.txt ./
+
+RUN pip install --upgrade pip && pip install --no-cache-dir -r requirements.txt
+
+COPY src ${LAMBDA_TASK_ROOT}/src
+
+CMD ["src.app.main.lambda_handler"]
