+++ fastAPI/NICKNAME-GENERATION-API/requirements.txt
