+++ fastAPI/NICKNAME-GENERATION-API/src/app/api/nickname/schema/nickname_profile_res.py
