+++ fastAPI/NICKNAME-GENERATION-API/src/app/api/nickname/schema/profileImage_req.py
