+++ fastAPI/NICKNAME-GENERATION-API/src/common/constants.py
@@ -1,0 +1,2 @@
+MIN_LENGTH = 2
+MAX_LENGTH = 15
