+++ fastAPI/NICKNAME-GENERATION-API/template.yaml
