--- fastAPI/dockerfile
@@ -1,9 +1,0 @@
-FROM public.ecr.aws/lambda/python:3.9
-
-COPY requirements.txt ./
-
-RUN pip install --upgrade pip && pip install --no-cache-dir -r requirements.txt
-
-COPY app ${LAMBDA_TASK_ROOT}/app
-
-CMD ["app.main.lambda_handler"]
